fix several warnings

+++ libs/ofxCv/include/ofxCv/Utilities.h
@@ -189,7 +189,7 @@ namespace ofxCv {
 	template <class T> inline ofPolyline toOf(const vector<cv::Point_<T> >& contour) {
 		ofPolyline polyline;
 		polyline.resize(contour.size());
-		for(int i = 0; i < contour.size(); i++) {
+		for(int i = 0; i < (int)contour.size(); i++) {
 			polyline[i].x = contour[i].x;
 			polyline[i].y = contour[i].y;
 		}
